// src/Models/ModelManagerModels/Dto.ts
export interface ModelVersionResponseDTO {
  modelVersionId: number;
  versionDescription?: string | undefined;
  modelName: string | undefined;
  defaultTestingMode?: string | undefined;
  //timestamp: Date;
  timestamp: string;
  userName?: string | undefined;
  capacity?: number;
  testCapacity?: number;
  nominalOutput?: number;
  nominalOutputPercentage?: number;
  nonlinearityPercentage?: number;
  minimumDeadLoad?: number;
  vMin?: number;
  nMax?: number;
  safeLoad?: number;
  ultimateLoad?: number;
  shellMaterialName?: string | undefined;
  versionHistory?: string | undefined;
  defaultCableLength?: number;
  resistance?: number;
  ccNumber?: string | undefined;
  accuracyClass?: string | undefined;
  application?: string | undefined;
  guidId?: string;

  compensatedTempRangeLow: number;
  compensatedTempRangeHigh: number;
  ratedExcitationMin: number;
  ratedExcitationMax: number;
  operatingTempMin: number;
  operatingTempMax: number;
  noOfGauges: number ;
  testpoints: DefaultTestPointResponseDTO[]; //| undefined;
}
// ...
export interface DefaultTestPointResponseDTO {
  modelVersionId: number;
  modelName: string | undefined;
  order: number;
  versionGuid: string;
  capacityDefaultTestPoint?: number;
  guidId: string;
}